requirements.txt
fastapi
uvicorn
fast_alpr
pillow
onnxruntime
python-multipart
numpy